+++ packages/insomnia/src/ui/components/websockets/event-log-view.tsx
@@ -6,7 +6,6 @@ import styled from 'styled-components';
 
 import { WebSocketEvent } from '../../../main/network/websocket';
 import { SvgIcon, SvgIconProps } from '../svg-icon';
-import { getEventType, printEventData } from './event-utils';
 
 const Timestamp: FC<{ time: Date | number }> = ({ time }) => {
   const date = format(time, 'HH:mm:ss');
@@ -115,10 +114,10 @@ const EventMessageCell = styled('div')({
 const getMessage = (event: WebSocketEvent): string => {
   switch (event.type) {
     case 'message': {
-      if (getEventType(event) === 'Binary') {
+      if ('data' in event && typeof event.data === 'object') {
         return 'Binary data';
       }
-      return printEventData(event);
+      return event.data.toString('utf-8');
     }
     case 'open': {
       return 'Connected successfully';
